archive/20251109_cleanup/requirements.txt
langchain>=0.1.0
langchain-community>=0.1.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdown>=3.5.0
python-frontmatter>=1.0.0
requests>=2.31.0
flask>=2.3.0
